:truck: Replace the original Bruneton model with the generalized one
